restore unrelated tests
