+++ sql/core/src/test/scala/org/apache/spark/sql/ApproxTopKSuite.scala
@@ -339,33 +339,6 @@ class ApproxTopKSuite extends QueryTest
     checkAnswer(res, Row(Seq(Row("b", 3), Row("a", 2))))
   }
 
-  test("SPARK-52626: Support group by Time column") {
-    val ts1 = "15:00:00"
-    val ts2 = "22:00:00"
-    val localTime = Seq(ts1, ts1, ts2).map(LocalTime.parse)
-    val df = localTime.toDF("t").groupBy("t").count().orderBy("t")
-    val expectedSchema =
-      new StructType().add(StructField("t", TimeType())).add("count", LongType, false)
-    assert(df.schema == expectedSchema)
-    checkAnswer(df, Seq(Row(LocalTime.parse(ts1), 2), Row(LocalTime.parse(ts2), 1)))
-  }
-
-  test("SPARK-52660: Support aggregation of Time column when codegen is split") {
-    val res = sql(
-      "SELECT max(expr), MIN(expr) " +
-        "FROM VALUES TIME'22:01:00', " +
-        "TIME'22:00:00', " +
-        "TIME'15:00:00', " +
-        "TIME'22:01:00', " +
-        "TIME'13:22:01', " +
-        "TIME'03:00:00', " +
-        "TIME'22:00:00', " +
-        "TIME'17:45:00' AS tab(expr);")
-    checkAnswer(
-      res,
-      Row(LocalTime.of(22, 1, 0), LocalTime.of(3, 0, 0)))
-  }
-
   test("SPARK-52588: accumulate and estimate of Integer with default parameters") {
     val res = sql("SELECT approx_top_k_estimate(approx_top_k_accumulate(expr)) " +
       "FROM VALUES (0), (0), (0), (1), (1), (2), (3), (4) AS tab(expr);")
